Fix a 😱 oversight: checking the wrong variable for emptyness!
This could have led to a use-after-move (depending on the usage on the callee side) and ill-formedness. In this case, by pure luck, it was benign but still wrong. The check was (also) there to protect its subsequent use.

Thanks Nathan for making me aware of this. How could have this escaped me for so long! 🤦‍♀️